Refactor: moved common/tts_model.py to style_bert_vits2/
